Major improvements and bug fixes for face detection and UI
- Fixed Kotlin version compatibility (1.9.20 for Compose Compiler 1.5.4)
- Fixed missing launcher icon resources (using Android default icon)
- Removed OpenCV dependency and replaced with Android bitmap processing
- Fixed memory issues with increased Gradle heap size (6GB)
- Fixed progress calculation to accurately reflect processing stages
- Made face detection much more lenient (0.01 min face size, no pixel restrictions)
- Added extensive debugging logs for face detection troubleshooting
- Added fallback demo mode when no faces are detected
- Fixed deprecated media scanner API for Android 11+
- Added real-time face count display in progress messages
- Improved error handling and user feedback
- Added FileProvider for sharing functionality
- Updated gradle.properties with build optimization settings

+++ app/src/main/java/com/ifykyk/facecollage/collage/CollageGenerator.kt
@@ -171,14 +171,14 @@ class CollageGenerator {
         
         // Draw background
         val labelX = x + (width - labelWidth) / 2
-        val labelRect = RectF(labelX, y.toFloat(), labelX + labelWidth, y + labelHeight)
+        val labelRect = RectF(labelX, y.toFloat(), labelX + labelWidth, (y + labelHeight))
         canvas.drawRoundRect(labelRect, 5f, 5f, backgroundPaint)
         
         // Draw title
         canvas.drawText(
             title,
             labelX + 10,
-            y + 20,
+            (y + 20).toFloat(),
             paint
         )
         
@@ -186,7 +186,7 @@ class CollageGenerator {
         canvas.drawText(
             subtitle,
             labelX + 10,
-            y + 40,
+            (y + 40).toFloat(),
             subtitlePaint
         )
     }
@@ -252,7 +252,7 @@ class CollageGenerator {
                 canvas.drawCircle(centerX, centerY, radius, paint)
                 
                 // Draw circular image
-                val scaledBitmap = scaleCenterCrop(representativeShot.bitmap, tileSize.toInt(), tileSize.toInt())
+                val scaledBitmap = scaleCenterCrop(representativeShot.bitmap, tileSize, tileSize)
                 val circularBitmap = createCircularBitmap(scaledBitmap)
                 canvas.drawBitmap(circularBitmap, x, y, paint)
